requirements.txt
langgraph==0.2.55
langchain-google-genai==2.0.7
langchain-core==0.3.28
pydantic==2.9.2
httpx==0.27.2
streamlit==1.39.0
python-dotenv==1.0.1
pytest==8.3.3
